pyproject.toml
[tool.poetry]
name = "slondrf"
version = "0.1.0"
description = "slon bot on drf"
authors = ["Psycho Asto <kuvms068@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
aiogram = "^3.4.1"
python-dotenv = "^1.0.1"
aiogram-dialog = "^2.1.0"
drf-spectacular = "^0.27.1"
django-cors-headers = "^4.3.1"
django-loki = "^0.1.4"
django-debug-toolbar = "^4.3.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
uvicorn = "^0.29.0"
python-json-logger = "^2.0.7"
django-asgi-lifespan = "^0.3.1"
pillow = "^10.3.0"
django = "^5.0.4"
django-jazzmin = "^2.6.1"
loguru = "^0.7.2"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.4.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
